Removed some barrel exports to reduce bundle size, moved some components to new folders, added toastify container
