+++ modules/PDFView/PDFInputs/Components/CertificatesInput.tsx
@@ -1,7 +1,11 @@
-import { useAnimation, usePDFData } from '@/Hooks';
 import { Operations } from '@/store/reducers/pdfData';
-import { AddNewButton, Certificate, DeleteButton } from '@modules/PDFView';
-import { DateInput, TextInput } from '@modules/Shared/Inputs';
+import { AddNewButton } from '@modules/PDFView/PDFInputs/Components/AddNewButton';
+import { DeleteButton } from '@modules/PDFView/PDFInputs/Components/DeleteButton';
+import { Certificate } from '@modules/PDFView/models';
+import { useAnimation } from '@modules/Shared/Hooks/useAnimation';
+import { usePDFData } from '@modules/Shared/Hooks/usePDFData';
+import { DateInput } from '@modules/Shared/Inputs/DateInput';
+import { TextInput } from '@modules/Shared/Inputs/TextInput';
 import { motion } from 'framer-motion';
 import { useTranslation } from 'next-i18next';
