+++ modules/PDFView/PDFInputs/Components/LanguagesInput.tsx
@@ -1,11 +1,10 @@
-import { useAnimation, usePDFData } from '@/Hooks';
 import { Operations } from '@/store/reducers/pdfData';
-import {
-  AddNewButton,
-  DeleteButton,
-  LanguageProficiencyLevel,
-} from '@modules/PDFView';
-import { TextInput } from '@modules/Shared/Inputs';
+import { AddNewButton } from '@modules/PDFView/PDFInputs/Components/AddNewButton';
+import { DeleteButton } from '@modules/PDFView/PDFInputs/Components/DeleteButton';
+import { LanguageProficiencyLevel } from '@modules/PDFView/models';
+import { useAnimation } from '@modules/Shared/Hooks/useAnimation';
+import { usePDFData } from '@modules/Shared/Hooks/usePDFData';
+import { TextInput } from '@modules/Shared/Inputs/TextInput';
 import { motion } from 'framer-motion';
 import { useTranslation } from 'next-i18next';
