+++ modules/PDFView/PDFInputs/Components/SkillsInput.tsx
@@ -1,7 +1,9 @@
-import { useAnimation, usePDFData } from '@/Hooks';
 import { Operations } from '@/store/reducers/pdfData';
-import { AddNewButton, SkillsList } from '@modules/PDFView/PDFInputs';
-import { TextInput } from '@modules/Shared/Inputs';
+import { AddNewButton } from '@modules/PDFView/PDFInputs/Components/AddNewButton';
+import { SkillsList } from '@modules/PDFView/PDFInputs/Components/SkillsList';
+import { useAnimation } from '@modules/Shared/Hooks/useAnimation';
+import { usePDFData } from '@modules/Shared/Hooks/usePDFData';
+import { TextInput } from '@modules/Shared/Inputs/TextInput';
 import { motion } from 'framer-motion';
 import { useTranslation } from 'next-i18next';
 import { useState } from 'react';
