--- modules/PDFView/PDFInputs/index.ts
@@ -1,3 +1,0 @@
-export * from './Components';
-export * from './PDFInputsContainer';
-export * from './PDFInputsPresenter';
