--- modules/PDFView/PDFTabNavigation/index.ts
@@ -1,1 +1,0 @@
-export * from './PDFTabNavigationPresenter';
