+++ modules/PDFView/index.ts
@@ -1,8 +1,4 @@
 export * from './CVMakerPage';
 export * from './CVTemplates';
 export * from './CreateView';
-export * from './PDFDownload';
-export * from './PDFInputs';
-export * from './PDFTabNavigation';
-export * from './VersionHistory';
 export * from './models';
