--- modules/Shared/Inputs/index.ts
@@ -1,3 +1,0 @@
-export * from './DateInput';
-export * from './TextInput';
-export * from './ToggleInput';
