+++ package.json
@@ -3,20 +3,20 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@react-pdf/renderer": "^3.1.5",
+    "@react-pdf/renderer": "^3.1.7",
     "@reduxjs/toolkit": "^1.9.3",
     "axios": "^1.3.4",
-    "compressorjs": "^1.1.1",
+    "compressorjs": "^1.2.1",
     "disable-scroll": "^0.6.0",
-    "firebase": "^9.17.1",
-    "framer-motion": "^9.0.7",
-    "i18next": "^22.4.10",
+    "firebase": "^9.17.2",
+    "framer-motion": "^10.2.3",
+    "i18next": "^22.4.11",
     "i18next-browser-languagedetector": "^7.0.1",
     "i18next-intervalplural-postprocessor": "^3.0.0",
     "js-cookie": "^3.0.1",
-    "luxon": "^3.2.1",
-    "next": "^13.1.6",
-    "next-i18next": "^13.1.6",
+    "luxon": "^3.3.0",
+    "next": "^13.2.3",
+    "next-i18next": "^13.2.1",
     "next-language-detector": "^1.0.2",
     "react": "^18.2.0",
     "react-custom-scroll": "^5.0.0",
@@ -63,7 +63,7 @@
   },
   "proxy": "http://localhost:3000",
   "devDependencies": {
-    "@next/bundle-analyzer": "^13.1.6",
+    "@next/bundle-analyzer": "^13.2.3",
     "@svgr/webpack": "^6.5.1",
     "@testing-library/jest-dom": "^5.16.5",
     "@testing-library/react": "^14.0.0",
@@ -71,7 +71,7 @@
     "@types/jest": "^29.4.0",
     "@types/js-cookie": "^3.0.3",
     "@types/luxon": "^3.2.0",
-    "@types/node": "^18.14.0",
+    "@types/node": "^18.14.6",
     "@types/react": "^18.0.28",
     "@types/react-custom-scroll": "^4.3.2",
     "@types/react-custom-scrollbars": "^4.0.10",
@@ -79,27 +79,27 @@
     "@types/react-pdf": "^6.2.0",
     "@types/webpack": "^5.28.0",
     "@types/webpack-env": "^1.18.0",
-    "@typescript-eslint/eslint-plugin": "^5.53.0",
-    "@typescript-eslint/parser": "^5.53.0",
+    "@typescript-eslint/eslint-plugin": "^5.54.1",
+    "@typescript-eslint/parser": "^5.54.1",
     "autoprefixer": "^10.4.13",
     "copy-webpack-plugin": "^11.0.0",
     "css-loader": "^6.7.3",
     "duplicate-dependencies-webpack-plugin": "^1.0.2",
-    "eslint": "^8.34.0",
+    "eslint": "^8.35.0",
     "eslint-plugin-react": "^7.32.2",
     "eslint-plugin-react-hooks": "^4.6.0",
-    "firebase-tools": "^11.23.1",
+    "firebase-tools": "^11.24.1",
     "fs": "^0.0.1-security",
     "html-webpack-plugin": "^5.5.0",
-    "mini-css-extract-plugin": "^2.7.2",
-    "nodemon": "^2.0.20",
+    "mini-css-extract-plugin": "^2.7.3",
+    "nodemon": "^2.0.21",
     "path": "^0.12.7",
     "postcss": "^8.4.21",
     "postcss-loader": "^7.0.2",
     "postcss-preset-env": "^8.0.1",
     "prettier": "^2.8.4",
     "prettier-plugin-css-order": "^1.3.0",
-    "prettier-plugin-tailwindcss": "^0.2.3",
+    "prettier-plugin-tailwindcss": "^0.2.4",
     "process": "^0.11.10",
     "sass": "^1.58.3",
     "sass-loader": "^13.2.0",
