--- src/store/index.ts
@@ -1,5 +1,0 @@
-export * from './actions';
-export * from './fbConfig-prod';
-export * from './fbConfig';
-export * from './hooks';
-export * from './store';
